+++ src/com/company/People/Person.java
@@ -3,11 +3,11 @@ package com.company.People;
 public abstract class Person {
     public String name;
     public String surname;
-    protected Double money;
+    protected Integer money;
 
     public Person(){
     }
-    public Person(String name, String surname,Double money){
+    public Person(String name, String surname,Integer money){
         this.name = name;
         this.surname = surname;
         this.money = money;
