+++ src/com/company/Vehicles/Van.java
@@ -3,7 +3,7 @@ package com.company.Vehicles;
 public class Van extends Car{
     public Double loadArea;
 
-    public Van(Double value, String mark, String model, Integer mileage,
+    public Van(Integer value, String mark, String model, Integer mileage,
                String color, String segment, Part[] parts, Double loadArea) {
         super(value, mark, model, mileage, color, segment, parts);
         this.loadArea = loadArea;
